+++ routes/sandwich.js
@@ -34,6 +34,9 @@ router.post("/", [auth, admin], async (req, res, next) => {
 
 router.put("/:id", [auth, admin] , async (req, res, next) => {
   try {
+    const isValidId = mongoose.isValidObjectId(req.params.id);
+    if (!isValidId)
+      return res.status(400).send("Invalid drink ID.");
     const { error } = validate(req.body);
     if (error) return res.status(400).send(error.details[0].message);
 
@@ -60,6 +63,9 @@ router.put("/:id", [auth, admin] , async (req, res, next) => {
 
 router.delete("/:id", [auth, admin],  async (req, res, next) => {
   try {
+    const isValidId = mongoose.isValidObjectId(req.params.id);
+    if (!isValidId)
+      return res.status(400).send("Invalid drink ID.");
     const sandwich = await Sandwich.findByIdAndRemove(req.params.id);
 
     if (!sandwich)
@@ -75,6 +81,9 @@ router.delete("/:id", [auth, admin],  async (req, res, next) => {
 
 router.get("/:id",  async (req, res, next) => {
   try {
+    const isValidId = mongoose.isValidObjectId(req.params.id);
+    if (!isValidId)
+      return res.status(400).send("Invalid drink ID.");
     const sandwich = await Sandwich.findById(req.params.id);
 
     if (!sandwich)
